code re-organization

+++ Entities/Entities.Lab/ParameterResultConfig.cs
@@ -28,35 +28,35 @@ namespace Entities.Lab
         /// <value>
         /// The minimum boundary.
         /// </value>
-        public Double? MinBoundary { get; set; }
+        public decimal? MinBoundary { get; set; }
         /// <summary>
         /// Gets or sets the maximum boundary.
         /// </summary>
         /// <value>
         /// The maximum boundary.
         /// </value>
-        public Double? MaxBoundary { get; set; }
+        public decimal? MaxBoundary { get; set; }
         /// <summary>
         /// Gets or sets the maximum normal range.
         /// </summary>
         /// <value>
         /// The maximum normal range.
         /// </value>
-        public Double? MaxNormalRange { get; set; }
+        public decimal? MaxNormalRange { get; set; }
         /// <summary>
         /// Gets or sets the minimum normal range.
         /// </summary>
         /// <value>
         /// The minimum normal range.
         /// </value>
-        public Double? MinNormalRange { get; set; }
+        public decimal? MinNormalRange { get; set; }
         /// <summary>
         /// Gets or sets the detection limit.
         /// </summary>
         /// <value>
         /// The detection limit.
         /// </value>
-        public Double? DetectionLimit { get; set; }
+        public decimal? DetectionLimit { get; set; }
         /// <summary>
         /// Gets the unit identifier.
         /// </summary>
